==> azure-ml/01_exploration.ipynb <==
# Used Car Price Prediction – Exploration

## Opis problema
Cilj je predvidjeti cijenu rabljenog automobila (*price*) na temelju njegovih značajki
(proizvođač, model, godina, kilometraža, gorivo, motor, mjenjač, boje, stanje, itd.).

## Skup podataka
Korišten je Kaggle dataset “Used Car Price Prediction Dataset” koji sadrži tablične podatke
o rabljenim automobilima. Problem je regresija jer je ciljna varijabla numerička.


import pandas as pd

df = pd.read_csv("used_cars.csv")
df.head()


df.shape

df.info()

df.isnull().sum()

df.describe(include='all').T


## Zaključak (EDA)
Skup podataka sadrži kombinaciju numeričkih i kategoričkih značajki.
Uočene su nedostajuće vrijednosti u određenim stupcima, a ciljna varijabla je `price`.
U sljedećem koraku provodi se čišćenje i predobrada (pretvorba `price` i `milage` u numerički oblik,
imputacija nedostajućih vrijednosti, kodiranje kategoričkih varijabli i normalizacija).
