==> azure-ml/02_preprocessing.ipynb <==
import pandas as pd

df = pd.read_csv("used_cars.csv")
df.head()

df['price'] = (
    df['price']
    .str.replace('$', '', regex=False)
    .str.replace(',', '', regex=False)
    .astype(float)
)


df['price'].head()

df['milage'] = (
    df['milage']
    .str.replace(' mi.', '', regex=False)
    .str.replace(',', '', regex=False)
    .astype(float)
)


df['milage'].head()


df['fuel_type'] = df['fuel_type'].fillna('Unknown')
df['accident'] = df['accident'].fillna('Unknown')
df['clean_title'] = df['clean_title'].fillna('Unknown')


df.isnull().sum()


X = df.drop(columns=['price'])
y = df['price']


from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42
)


categorical_features = X.select_dtypes(include=['object']).columns.tolist()
numerical_features = X.select_dtypes(include=['int64', 'float64']).columns.tolist()

categorical_features, numerical_features


from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline


numeric_transformer = Pipeline(steps=[
    ('scaler', StandardScaler())
])


categorical_transformer = Pipeline(steps=[
    ('onehot', OneHotEncoder(handle_unknown='ignore'))
])


preprocessor = ColumnTransformer(
    transformers=[
        ('num', numeric_transformer, numerical_features),
        ('cat', categorical_transformer, categorical_features)
    ]
)


X_train_processed = preprocessor.fit_transform(X_train)
X_test_processed = preprocessor.transform(X_test)


X_train_processed.shape, X_test_processed.shape


## Kodiranje i normalizacija podataka

Numeričke značajke (*model_year*, *milage*) normalizirane su korištenjem
StandardScaler metode kako bi se osigurala usporedivost vrijednosti.

Kategoričke značajke kodirane su primjenom One-Hot Encoding tehnike.
Za obradu potencijalno nepoznatih kategorija u testnim i produkcijskim
podacima korištena je opcija *handle_unknown='ignore'*.

Za objedinjavanje svih koraka predobrade korišten je *ColumnTransformer*,
čime je osigurana konzistentna primjena predobrade tijekom treniranja
i kasnije prilikom deploya modela.


import joblib

joblib.dump(preprocessor, "preprocessor.pkl")
