==> azure-ml/06_deploy_endpoint.ipynb <==
from azure.ai.ml import MLClient
from azure.identity import DefaultAzureCredential

ml_client = MLClient.from_config(DefaultAzureCredential())
ml_client


from azure.ai.ml.entities import Model

model_asset = Model(
    path="model_package",
    name="used-car-model-package",
    description="RandomForest model + preprocessor (OHE+Scaler) for used car price prediction",
    type="custom_model"
)

model_asset = ml_client.models.create_or_update(model_asset)
model_asset.name, model_asset.version


import time
from azure.ai.ml.entities import ManagedOnlineEndpoint, ManagedOnlineDeployment, Environment, CodeConfiguration
from azure.ai.ml.constants import AssetTypes

endpoint_name = "usedcar-endpoint-" + str(int(time.time()))
endpoint = ManagedOnlineEndpoint(
    name=endpoint_name,
    description="Used car price prediction endpoint"
)

ml_client.online_endpoints.begin_create_or_update(endpoint).result()
endpoint_name


# obriši deployment (ako postoji)
ml_client.online_deployments.begin_delete(
    name="blue",
    endpoint_name=endpoint_name
).result()


from azure.ai.ml.entities import Environment

env = Environment(
    name="used-car-env",
    description="Conda env for used car scoring",
    conda_file="azure_endpoint/conda.yaml",
    image="mcr.microsoft.com/azureml/minimal-ubuntu20.04-py38-cpu-inference:latest"
)


from azure.ai.ml import MLClient
from azure.identity import DefaultAzureCredential

ml_client = MLClient.from_config(DefaultAzureCredential())
ml_client


endpoint_name = "usedcar-endpoint-1770148704"


print(
    ml_client.online_deployments.get_logs(
        name="blue",
        endpoint_name=endpoint_name,
        lines=2000
    )
)


model_asset = ml_client.models.get(name="used-car-model-package", label="latest")
model_asset


from azure.ai.ml.entities import Environment

aml_env = Environment(
    name="used-car-env",
    description="Conda env for used car scoring",
    conda_file="azure_endpoint/conda.yaml",
    image="mcr.microsoft.com/azureml/minimal-ubuntu20.04-py38-cpu-inference:latest"
)


from azure.ai.ml.entities import ManagedOnlineDeployment, CodeConfiguration

deployment = ManagedOnlineDeployment(
    name="blue",
    endpoint_name=endpoint_name,
    model=model_asset,
    environment=aml_env,
    code_configuration=CodeConfiguration(code="azure_endpoint", scoring_script="score.py"),
    instance_type="Standard_DS1_v2",
    instance_count=1,
)

ml_client.online_deployments.begin_create_or_update(deployment).result()


endpoint = ml_client.online_endpoints.get(endpoint_name)
endpoint.traffic = {"blue": 100}
ml_client.online_endpoints.begin_create_or_update(endpoint).result()


endpoint = ml_client.online_endpoints.get(endpoint_name)
keys = ml_client.online_endpoints.get_keys(endpoint_name)

scoring_uri = endpoint.scoring_uri
api_key = keys.primary_key

scoring_uri


import requests, json

sample = {
    "brand": "Ford",
    "model": "Utility Police Interceptor Base",
    "model_year": 2013,
    "milage": 51000,
    "fuel_type": "E85 Flex Fuel",
    "engine": "300.0HP 3.7L V6 Cylinder Engine Flex Fuel Capability",
    "transmission": "6-Speed A/T",
    "ext_col": "Black",
    "int_col": "Black",
    "accident": "At least 1 accident or damage reported",
    "clean_title": "Yes"
}

headers = {
    "Content-Type": "application/json",
    "Authorization": f"Bearer {api_key}"
}

r = requests.post(scoring_uri, headers=headers, data=json.dumps(sample))
print(r.status_code)
print(r.text)
